+++ GitHub/Steps/PublishPowerShellGallery.ps1
@@ -1,15 +1,29 @@
 ﻿param(
 [string]
-$ModulePath
+$ModulePath,
+
+[string[]]
+$Exclude = @('*.png', '*.mp4', '*.jpg','*.jpeg', '*.gif', 'docs[/\]*')
 )
+
 $gitHubEvent = if ($env:GITHUB_EVENT_PATH) {
     [IO.File]::ReadAllText($env:GITHUB_EVENT_PATH) | ConvertFrom-Json
 } else { $null }
+
+if (-not $Exclude) {
+    $Exclude = @('*.png', '*.mp4', '*.jpg','*.jpeg', '*.gif','docs[/\]*')
+}
 
 
 @"
 ::group::GitHubEvent
 $($gitHubEvent | ConvertTo-Json -Depth 100)
+::endgroup::
+"@ | Out-Host
+
+@"
+::group::PSBoundParameters
+$($PSBoundParameters | ConvertTo-Json -Depth 100)
 ::endgroup::
 "@ | Out-Host
 
@@ -30,9 +44,9 @@ if (-not $ModulePath) {
 
 if (-not $imported) { return } 
 
-$foundModule = try { Find-Module -Name $imported.Name -ErrorAction SilentlyContinue } catch {}
+$foundModule = try { Find-Module -Name $imported.Name -ErrorAction SilentlyContinue} catch {}
 
-if ($foundModule -and $foundModule.Version -ge $imported.Version) {
+if ($foundModule -and (([Version]$foundModule.Version) -ge ([Version]$imported.Version))) {
     "::warning::Gallery Version of $moduleName is more recent ($($foundModule.Version) >= $($imported.Version))" | Out-Host        
 } else {
     
@@ -55,9 +69,24 @@ if ($foundModule -and $foundModule.Version -ge $imported.Version) {
     if (Test-Path $moduleGitPath) {
         Remove-Item -Recurse -Force $moduleGitPath
     }
+
+    if ($Exclude) {
+        "::notice::Attempting to Exlcude $exclude" | Out-Host
+        Get-ChildItem $moduleTempPath -Recurse |
+            Where-Object {                
+                foreach ($ex in $exclude) {
+                    if ($_.FullName -like $ex) {
+                        "::notice::Excluding $($_.FullName)" | Out-Host
+                        return $true
+                    }
+                }
+            } | 
+            Remove-Item
+    }
+
     Write-Host "Module Files:"
     Get-ChildItem $moduleTempPath -Recurse
-    Write-Host "Publishing $moduleName [$($imported.Version)] to Gallery"
+    Write-Host "Publishing $moduleName [$($imported.Version)] to Gallery"    
     Publish-Module -Path $moduleTempPath -NuGetApiKey $gk
     if ($?) {
         Write-Host "Published to Gallery"