+++ PSDevOps.ezformat.ps1
@@ -1,52 +1,39 @@
-﻿#requires -Module EZOut
-# Install-Module EZOut or https://github.com/StartAutomating/EZOut
-[Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSAvoidUsingWriteHost", "", Justification="This generates format files (where its ok to Write-Host)")]
-param()
+#requires -Module EZOut
+#  Install-Module EZOut or https://github.com/StartAutomating/EZOut
 $myFile = $MyInvocation.MyCommand.ScriptBlock.File
-$myModuleName = $($myFile | Split-Path -Leaf) -replace '\.ezformat\.ps1', ''
+$myModuleName = 'PSDevOps'
 $myRoot = $myFile | Split-Path
-
-
+Push-Location $myRoot
 $formatting = @(
-    Import-FormatView -FilePath (Join-Path $myRoot Formatting)
+    # Add your own Write-FormatView here,
+    # or put them in a Formatting or Views directory
+    foreach ($potentialDirectory in 'Formatting','Views') {
+        Join-Path $myRoot $potentialDirectory |
+            Get-ChildItem -ea ignore |
+            Import-FormatView -FilePath {$_.Fullname}
+    }
 )
 
-$myFormatFile = Join-Path $myRoot "$myModuleName.format.ps1xml"
-$formatting | Out-FormatData -ModuleName PSDevOps | Set-Content $myFormatFile -Encoding UTF8
+$destinationRoot = $myRoot
+
+if ($formatting) {
+    $myFormatFile = Join-Path $destinationRoot "$myModuleName.format.ps1xml"
+    $formatting | Out-FormatData -Module $MyModuleName | Set-Content $myFormatFile -Encoding UTF8
+    Get-Item $myFormatFile
+}
 
 $types = @(
-    <#
-    Write-TypeView -TypeName PSDevOps.ArtifactFeed.View -AliasProperty @{
-        ViewID = 'id'
-    }
-    Write-TypeView -TypeName PSDevOps.ArtifactFeed -AliasProperty @{
-        FeedID = 'fullyQualifiedId'
-    } -HideProperty ViewID
-    Write-TypeView -TypeName PSDevOps.Build.Definition -AliasProperty @{
-        DefinitionID = 'ID'
-    }
-    Write-TypeView -TypeName PSDevOps.Repository -AliasProperty @{
-        RepositoryID = 'ID'
-    }
-    Write-TypeView -TypeName PSDevOps.ServiceEndpoint -AliasProperty @{
-        EndpointID = 'ID'
-        EndpointType = 'Type'
-    }
-    Write-TypeView -TypeName PSDevOps.Repository.SourceProvider -AliasProperty @{
-        ProviderName = 'Name'
-    }
-    Write-TypeView -TypeName PSDevOps.Project -AliasProperty @{
-        Project = 'Name'
-        ProjectID = 'ID'
-    } -ScriptProperty @{
-        LastUpdated = {[DateTime]$this.LastUpdateTime}
-    } -DefaultDisplay Organization, Project, LastUpdateTime, Description
-    #>
+    # Add your own Write-TypeView statements here
+    # or declare them in the 'Types' directory
     Join-Path $myRoot Types |
         Get-Item -ea ignore |
         Import-TypeView -Deserialized
 
 )
 
-$myTypesFile = Join-Path $myRoot "$myModuleName.types.ps1xml"
-$types | Out-TypeData | Set-Content $myTypesFile -Encoding UTF8
+if ($types) {
+    $myTypesFile = Join-Path $destinationRoot "$myModuleName.types.ps1xml"
+    $types | Out-TypeData | Set-Content $myTypesFile -Encoding UTF8
+    Get-Item $myTypesFile
+}
+Pop-Location