+++ CHANGELOG.md
@@ -1,4 +1,22 @@
-﻿## 0.5.8:
+﻿## 0.5.9:
+* New-GitHubAction:  Adding -OutputPath (#162)
+* New-GitHubWorkflow:  Adding -OutputPath (#163)
+* New-ADOPipeline:  Adding -OutputPath (#164)
+* Updating Action (Preferring local bits and other cleanup) (#165)
+* Import-BuildStep: Adding -SourceFile and -BuildStepName (Fixes #153)
+* Updating ReleaseModule Step:  Allowing -ReleaseAsset (Fixes #160)
+* PublishPowerShellGallery :  Adding -Exclude (Fixes #159).  Fixing Module Version Comparison (Fixes #149)
+* Connect-ADO:  Adding -NoCache (Fixes #150)
+* Convert-BuildStep:  Fixing Github object parameter behavior (Fixes #158)
+* Adding Get-ADOServiceHealth (Fixes #152)
+* Adding Get-ADOAuditLog (Fixes #151)
+* Adding PSDevOps.PSSVG.ps1 (Fixes #157)
+* Updating Workflow (Fixes #155 Fixes #156)
+* Adding GitHub/On/Issue (Fixes #154)
+
+---
+
+## 0.5.8:
 * Running EZOut in Workflow (#148)
 * Adding support for HelpOut (#147)
 * Updating action (pushing changes) #144
@@ -13,6 +31,7 @@
 * Get-ADOBuild:  Passing along ProjectID and ProjectName
 * Get/Set-ADOPermission:  Repov2 issue (#140)
 * Recategorizing Functions
+
 ---
 
 ## 0.5.7:
@@ -21,6 +40,7 @@
 * Get/Set-ADOPermission:  Support for ServiceEndpoints (#137)
 * Set-ADOPermission:  Exposing specialized parameter sets (#138)
 * PSDevOps.WorkProcess objects now return .ProcessName and .ProcessID as alias properties
+
 ---
 
 ## 0.5.6:
@@ -34,12 +54,15 @@
 * Adding "On" files for issue creation, deletion, or modification (Fixes #132)
 * Adding "On" files for common scheduling needs (Fixes #134)
 * ReleaseNameFormat can not be customized in ReleaseModule step (Fixes #130)
+
 ---
 
 ## 0.5.5.1:
 * Allowing for [Reflection.AssemblyMetaData] attributes on step .ps1 files (to customize YAML).  Issue #123
 * Better support for Object and array parameters in Azure DevOps. Issue #125.
+
 ---
+
 ## 0.5.5:
 * Azure DevOps: Adding support for Shared Queries (Fixes #117)
  * Get-ADOWorkItem -SharedQuery can get shared queries
@@ -49,14 +72,20 @@
  * Adding Job/Step definitions to Release Module
  * Adding -NoCoverage to RunPester Step
  * Creating Example workflow that publishes PSDevOps.
+
 ---
+
 ## 0.5.4.2:
 * Adding Register-ADOArtifactFeed (Fixes #118)
+
 ---
+
 ## 0.5.4.1:
 * Fixing Invoke-ADORestApi issues:  #111,#114,#115
 * Attaching .BuildID and .DefinitionID properties to Get-ADOBuild where appropriate.
+
 ---
+
 ## 0.5.4:
 * Formatting Improvments:
  * Get-ADOField now includes .Type
@@ -74,11 +103,15 @@
   * TagModuleVersion
 * New-GitHubWorkflow/New-ADOPipeline now support -RootDirectory
 * Fixing pluralization / list issue with multiple GitHub Workflow "On"
+
 ---
+
 ## 0.5.3
 * Get-ADORepository  :  Adding -PullRequestID
 * New/Set-ADOWorkItem:  Fixing pipelining issue
+
 ---
+
 ## 0.5.2
 * Get-ADOTeam:  Adding alias -AreaPath for -TeamFieldValue, carrying on team property
 * Set-ADOTeam:  Support for -DefaultAreaPath/-AreaPath (TeamFieldValues api, fixing issue #92)
@@ -93,12 +126,16 @@
   * Improving Formatting of Work Items (#82)
   * Adding -Tag
 * Invoke-ADORestAPI:  Fixing issue with -QueryParameter
+
 ---
+
 ## 0.5.1
 * Bugfixes:
  * Get-ADOTest:  Fixing parameter sets and adding formatting.
  * Invoke-GitHubRESTAPI:  Only using .ContentEncoding when present.
+
 ---
+
 ## 0.5
 * Improved Git Functionality
  * New-GitHubAction
@@ -117,7 +154,9 @@
  * Get-ADOTest [new]
  * New-ADOPipeline now has -RootDirectory
  * Tons of New Type Definitions and Formatters
- ---
+
+---
+
 ## 0.4.9
 * New Command: Wait-ADOBuild
 * Start-ADOBuild
@@ -129,6 +168,7 @@
  * PSDevOps now includes a file to generate it's own build
  * PublishTest/CodeCoverage Results steps will always() run
  * Convert-BuildStep will add a .Name to each script step.
+
 ---
 
 ## 0.4.8
@@ -141,7 +181,9 @@
  * Get-ADOProject now has -TestRun, -TestPlan, -Release, and -PendingApproval (and better progress bars)
  * Get-ADOWorkItemType now has -Field
  * Commands for Picklists:  Add/Get/Remove/Update-ADOPicklist
+
 ---
+
 ## 0.4.7
 * New Commands:
  * Add/Get/Remove-ADOWiki
@@ -149,7 +191,9 @@
 * Bugfixes:
  * Honoring Get-ADOBuild -DefinitionName
 * Disconnect-ADO is now run prior at the start of Connect-ADO, and on module unload.
-----
+
+---
+
 ## 0.4.6
 * New-ADOPipeline/New-GitHubWorkflow:  Adding -BuildScript
 * Connect-ADO:  Auto-detecting connected user's teams and adding tab completion for -Project/-ProjectID/-Team/-TeamID
@@ -160,7 +204,9 @@
 * Get-ADOBuild:  Adding -DefinitionName
 * Invoke-ADORestAPI:  Passing content length of 0 when body is empty.
 * Updating README
+
 ---
+
 ## 0.4.5
 * New Commands:
  * Connect/Disconnect-ADO:  Caching connection info and saving default parameters!
@@ -174,14 +220,18 @@
 * Updated Commands:
   * Get-ADOUser/Get-ADOTeam : Additional Graph scenarios added.
   * New/Set-ADOWorkItem     : Added -BypassRule, -ValidateOnly, -SkipNotification
+
 ---
+
 ## 0.4.4
 * Get-ADOTask:  Adding -YAMLSchema.
 * Get-ADOTeam:  Adding -Setting/-FieldValue/-Iteration/-Board.
 * Get-ADOAreaPath/ADOIterationPath:  Making parameter names match cmdlet.  Honoring -AreaPath/-IterationPath.
 * Get-ADOProject:  Adding -ProcessConfiguration, -Plan, -PlanID, and -DeliveryTimeline.
 * New Command:  Set-ADOTeam
+
 ---
+
 ## 0.4.3
 * Renaming commands:
  * New-ADODashboard -> Add-ADODashboard
@@ -191,7 +241,9 @@
 * New Command: Get-ADOUser
 * Get-ADOAgentPool:  Fixed pipelining bug, added -AgentName/-IncludeCapability/-IncludeLastCompletedRequest/-IncludeAssignedRequest.
 * Set-ADOProject:  Can now -EnableFeature and -DisableFeature
+
 ---
+
 ## 0.4.2
 * Build Step Improvements:
  * New-ADOPipeline now has -PowerShellCore and -WindowsPowerShell
@@ -204,7 +256,9 @@
 * New Extension Commands: Enable/Disbale-ADOExtension
 * Improved formatting/types for Extensions.
 * Breaking change: Install/Uninstall-ADOExtension now accept -PublisherID and -ExtensionID, not -PublisherName and -ExtensionName.
+
 ---
+
 ## 0.4.1
 * More GitHub Functionality:
  * Write-GitHubDebug
@@ -219,7 +273,9 @@
  * Set-ADOProject
  * Repositories returned from a build definition are now decorated as PSDevOps.Repository
 * Improved testing and static analysis compliance
+
 ---
+
 ## 0.4
 * Overhaul of GitHub Workflow functionality.
  * New-GitHubAction renamed to New-GitHubWorkflow
@@ -228,54 +284,76 @@
  * Added Write-GitHubError/GitHubWarning (updating Write-ADOError/Write-ADOWarning for consistency)
  * Cleaning up GitHub Workflow parts
 * Get-ADOTask no longer has -ApiVersion parameter
+
 ---
+
 ## 0.3.9
 * New/Get/Remove-ADOWorkItemType:  Create/get/remove work custom work item types, states, rules, and behaviors.
 * Added Get-ADOBuild -IncludeAllProperty/-IncludeLatestBuild.
 * ScriptCop Integration:  PowerShelllStaticAnalysis stage now runs ScriptCop as well.
 * Improved ScriptAnalyzer Integration: Rule name is now outputted.
+
 ---
+
 ## 0.3.8
 * Add/Remove-ADOAreaPath
 * Formatter for AreaPaths
+
 ---
+
 ## 0.3.7
 * Convert-ADOPipeline now has -Passthru and -Wherefore
 * Get-ADOWorkProcess now has -Behavior and -WorkItemType
 * Get-ADOWorkItem now has -Mine, -CurrentIteration, -Comment, -Update, -Revision.
+
 ---
+
 ## 0.3.6.1
 * Convert-ADOPipeline now binds to .Variables property
 * Fixing bug in PSDevOps.WorkItem types file, which displayed in formatting.
+
 ---
+
 ## 0.3.6
 * Added: Get-ADOTask, Convert-ADOPipeline
+
 ---
+
 ## 0.3.5
 * New Command: Get-ADOTeam
 * Get-ADOBuild -CodeCoverage
 * Progress bars on Get-ADORepository
 * Slight refactoring to make progress bars easier in any function
+
 ---
+
 ## 0.3.4.1
 * Removing supplied parameters in commands generated by Import-ADOProxy.
+
 ---
+
 ## 0.3.4
 * New capability: Import-ADOProxy (Import a proxy module with for your ADO / TFS instance)
 * New REST Commands: Get-ADOAreaPath, Get-ADOIterationPath, Get-ADOExtension
 * More Features: Get-ADORepository -FileList
 * Massive Internal Refactoring (switching to dynamic parameters for -PersonalAccessToken etc, standardizing pstypenames)
+
 ---
+
 ## 0.3.3
 * Now Caching Personal Access Tokens!
 * URLEncoding all segments in Parts/ReplaceRouteParameter.
 * Ensuring all Azure DevOps YAML Parameters are wrapped in a string.
+
 ---
+
 ## 0.3.2
 * Pester workarounds - Steps/InstallPester and Steps/RunPester now accept a PesterMaxVersion (defaulting to 4.99.99)
 * Convert-BuildStep handles blank parameter defaults correctly
 * Get-ADOBuild can get yaml definitions directly, e.g Get-ADOBuild -DefinitionID 123 -DefinitionYaml
+
 ---
+
 ## 0.3.1
 * Bugfixes and Improvements to Convert/Import/Expand-BuildStep:
 1. Enforcing pluralization of certain fields within Azure DevOps
@@ -285,29 +363,39 @@
 * Added Convert/Import/Expand-BuildStep
 * Allowing build steps to be defined in functions
 * Automagically importing build step parameters
+
 ---
+
 ## 0.2.9
 * Get/New/Remove-ADORepository
 * Get/New/Remove-ADOServiceEndpoint
 * Get-ADOAgentPool
  * Improvements to New-ADOPipeline to avoid unexpected singletons
+
 ---
+
 ## 0.2.8
 * Get/New/Update-ADOBuild
 * Improving New-ADOPipeline:
  * Unknown -InputObject properties will no longer be pluralized
  * Added 'Pool' to list of known singletons
+
 ---
+
 ## 0.2.7
 * New Cmdlet: Set-ADOArtifactFeed
 * Improvements to New/Get/Remove-ADOArtifactFeed (better pipelining, renaming -FullyQualifiedID to -FeedID)
+
 ---
+
 ## 0.2.6
 * New Cmdlets:
  * New/Get/Remove-ADOArtifactFeed
  * New/Remove-ADOProject
  * New-GitHubAction
+
 ---
+
 ## 0.2.5
 * Improving Get-ADOWorkItem:
  * -Title allows getting work items by title
@@ -318,15 +406,22 @@
  * Formatting improved
 * Adding Get-ADOWorkProcess
 * Fixing issues with -ADOField commands when not provided a -Project
+
+---
+
 ## 0.2.4
 * Adding Adding -CanSortBy, -IsQueryable, and -ReadOnly to New-ADOField.
 * Adding parameter help to New-ADOField
+
 ---
+
 ## 0.2.3
 * Adding New/Remove-ADOField
 * Adding help to Get-ADOField
 * Adding formatting for fields
+
 ---
+
 ## 0.2.2
 * Adding New/Set/Remove-ADOWorkItem
 * Adding Get-ADOField
@@ -334,13 +429,19 @@
 * New Parameter: New-ADOPipeline -Option
 * Initial formatting
 * Switching Parts to use latest VMImage
+
 ---
+
 ## 0.2.1
 * Added Get-ADOWorkItem
+
 ---
+
 ## 0.2
 * Added Invoke-ADORestAPI
+
 ---
+
 ## 0.1
 Initial Commit
 ---